+++ next-frontend/src/globals/Privacy.ts
